clean Att+ Positional encodings layers
