--- faenet/layers.py
@@ -1,221 +1,0 @@
-import math
-from typing import Optional, Tuple, Union
-
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-from torch import Tensor
-from torch_geometric.nn.conv import MessagePassing
-from torch_geometric.nn.dense.linear import Linear
-from torch_geometric.typing import (
-    Adj,
-    OptTensor,
-    PairTensor,
-)
-from torch_geometric.utils import softmax
-from torch_sparse import SparseTensor
-
-
-class GaussianSmearing(nn.Module):
-    def __init__(self, start=0.0, stop=5.0, num_gaussians=50):
-        super().__init__()
-        offset = torch.linspace(start, stop, num_gaussians)
-        self.coeff = -0.5 / (offset[1] - offset[0]).item() ** 2
-        self.register_buffer("offset", offset)
-
-    def forward(self, dist):
-        dist = dist.view(-1, 1) - self.offset.view(1, -1)
-        return torch.exp(self.coeff * torch.pow(dist, 2))
-
-
-class PositionalEncoding(nn.Module):
-    def __init__(self, d_model: int, max_count: int = 210):
-        super().__init__()
-        position = torch.arange(max_count).unsqueeze(1)
-        div_term = torch.exp(
-            torch.arange(0, d_model, 2) * (-math.log(10000.0) / d_model)
-        )
-        pe = torch.zeros(max_count, 1, d_model)
-        pe[:, 0, 0::2] = torch.sin(position * div_term)
-        pe[:, 0, 1::2] = torch.cos(position * div_term)
-        pe = pe.squeeze(1)
-        self.register_buffer("pe", pe)
-
-    def forward(self, x: Tensor) -> Tensor:
-        """
-        Args:
-            x: Tensor, shape [num_atoms, batch_size, embedding_dim]
-            x: Tensor, shape [num_atoms]
-        """
-        return self.pe[x]
-
-
-class TransfoAttConv(MessagePassing):  # TODO: vient d'où?
-    r"""The graph transformer operator from the `"Masked Label Prediction:
-    Unified Message Passing Model for Semi-Supervised Classification"
-    <https://arxiv.org/abs/2009.03509>`_ paper
-
-    Args:
-        in_channels (int or tuple): Size of each input sample, or :obj:`-1` to
-            derive the size from the first input(s) to the forward method.
-            A tuple corresponds to the sizes of source and target
-            dimensionalities.
-        out_channels (int): Size of each output sample.
-        heads (int, optional): Number of multi-head-attentions.
-            (default: :obj:`1`)
-        concat (bool, optional): If set to :obj:`False`, the multi-head
-            attentions are averaged instead of concatenated.
-            (default: :obj:`True`)
-        beta (bool, optional): If set, will combine aggregation and
-            skip information via
-        dropout (float, optional): Dropout probability of the normalized
-            attention coefficients which exposes each node to a stochastically
-            sampled neighborhood during training. (default: :obj:`0`)
-        edge_dim (int, optional): Edge feature dimensionality (in case
-            there are any). Edge features are convoled with value features prior
-            multiplication by attention coefficient
-        bias (bool, optional): If set to :obj:`False`, the layer will not learn
-            an additive bias. (default: :obj:`True`)
-        root_weight (bool, optional): If set to :obj:`False`, the layer will
-            not add the transformed root node features to the output and the
-            option  :attr:`beta` is set to :obj:`False`. (default: :obj:`True`)
-        **kwargs (optional): Additional arguments of
-            :class:`torch_geometric.nn.conv.MessagePassing`.
-    """
-    _alpha: OptTensor
-
-    def __init__(
-        self,
-        in_channels: Union[int, Tuple[int, int]],
-        out_channels: int,
-        heads: int = 1,
-        concat: bool = True,
-        dropout: float = 0.0,
-        edge_dim: Optional[int] = None,
-        bias: bool = True,
-        root_weight: bool = True,
-        **kwargs,
-    ):
-        kwargs.setdefault("aggr", "add")
-        super(TransfoAttConv, self).__init__(node_dim=0, **kwargs)
-
-        self.in_channels = in_channels
-        self.out_channels = out_channels
-        self.heads = heads
-        self.root_weight = root_weight
-        self.concat = concat
-        self.dropout = dropout
-        self.edge_dim = edge_dim
-        self._alpha = None
-
-        if isinstance(in_channels, int):
-            in_channels = (in_channels, in_channels)
-
-        self.lin_key = Linear(in_channels[0], heads * out_channels)
-        self.lin_query = Linear(in_channels[1], heads * out_channels)
-        self.lin_value = Linear(in_channels[0], heads * out_channels)
-        if edge_dim is not None:
-            self.lin_edge = Linear(edge_dim, heads * out_channels, bias=False)
-        else:
-            self.lin_edge = self.register_parameter("lin_edge", None)
-
-        if concat:
-            self.lin_skip = Linear(in_channels[1], heads * out_channels, bias=bias)
-        else:
-            self.lin_skip = Linear(in_channels[1], out_channels, bias=bias)
-
-        self.reset_parameters()
-
-    def reset_parameters(self):
-        self.lin_key.reset_parameters()
-        self.lin_query.reset_parameters()
-        self.lin_value.reset_parameters()
-        if self.edge_dim:
-            self.lin_edge.reset_parameters()
-        self.lin_skip.reset_parameters()
-
-    def forward(
-        self,
-        x: Union[Tensor, PairTensor],
-        edge_index: Adj,
-        edge_attr: OptTensor = None,
-        return_attention_weights=None,
-    ):
-        r"""
-        Args:
-            return_attention_weights (bool, optional): If set to :obj:`True`,
-                will additionally return the tuple
-                :obj:`(edge_index, attention_weights)`, holding the computed
-                attention weights for each edge. (default: :obj:`None`)
-        """
-
-        H, C = self.heads, self.out_channels
-
-        if isinstance(x, Tensor):
-            x: PairTensor = (x, x)
-
-        query = self.lin_query(x[1]).view(-1, H, C)
-        key = self.lin_key(x[0]).view(-1, H, C)
-        value = self.lin_value(x[0]).view(-1, H, C)
-
-        # propagate_type: (query: Tensor, key:Tensor, value: Tensor, edge_attr: OptTensor) # noqa
-        out = self.propagate(
-            edge_index,
-            query=query,
-            key=key,
-            value=value,
-            edge_attr=edge_attr,
-            size=None,
-        )
-
-        alpha = self._alpha
-        self._alpha = None
-
-        if self.concat:
-            out = out.view(-1, self.heads * self.out_channels)
-        else:
-            out = out.mean(dim=1)
-
-        if self.root_weight:
-            x_r = self.lin_skip(x[1])
-            out = out + x_r
-
-        if isinstance(return_attention_weights, bool):
-            assert alpha is not None
-            if isinstance(edge_index, Tensor):
-                return out, (edge_index, alpha)
-            elif isinstance(edge_index, SparseTensor):
-                return out, edge_index.set_value(alpha, layout="coo")
-        else:
-            return out
-
-    def message(
-        self,
-        query_i: Tensor,
-        key_j: Tensor,
-        value_j: Tensor,
-        edge_attr: OptTensor,
-        index: Tensor,
-        ptr: OptTensor,
-        size_i: Optional[int],
-    ) -> Tensor:
-
-        # Compute edge embed
-        if self.lin_edge is not None:
-            assert edge_attr is not None
-            edge_attr = self.lin_edge(edge_attr).view(-1, self.heads, self.out_channels)
-
-        # Compute attention coefficient
-        alpha = (query_i * key_j).sum(dim=-1) / math.sqrt(self.out_channels)
-        alpha = softmax(alpha, index, ptr, size_i)
-        self._alpha = alpha
-        alpha = F.dropout(alpha, p=self.dropout, training=self.training)
-
-        out = value_j * alpha.view(-1, self.heads, 1) * edge_attr
-        return out
-
-    def __repr__(self) -> str:
-        return (
-            f"{self.__class__.__name__}({self.in_channels}, "
-            f"{self.out_channels}, heads={self.heads})"
-        )
