+++ faenet/model.py
@@ -14,8 +14,7 @@ from torch_scatter import scatter
 from faenet.base_model import BaseModel
 from faenet.embedding import PhysEmbedding
 from faenet.force_decoder import ForceDecoder
-from faenet.layers import GaussianSmearing
-from faenet.utils import get_pbc_distances
+from faenet.utils import get_pbc_distances, GaussianSmearing
 
 try:
     from torch_geometric.nn.acts import swish
