+++ faenet/utils.py
@@ -1,4 +1,5 @@
 import torch
+import torch.nn as nn
 
 
 def get_pbc_distances(
@@ -40,3 +41,14 @@ def get_pbc_distances(
         out["offsets"] = offsets[nonzero_idx]
 
     return out
+
+class GaussianSmearing(nn.Module):
+    def __init__(self, start=0.0, stop=5.0, num_gaussians=50):
+        super().__init__()
+        offset = torch.linspace(start, stop, num_gaussians)
+        self.coeff = -0.5 / (offset[1] - offset[0]).item() ** 2
+        self.register_buffer("offset", offset)
+
+    def forward(self, dist):
+        dist = dist.view(-1, 1) - self.offset.view(1, -1)
+        return torch.exp(self.coeff * torch.pow(dist, 2))
